Moving money spring
Both money-spring3 and money-spring are moved to use the new money core
